Rename utilities module to follow Django convention
